+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changes by version
 
+* added new credentials types: `SaKeyJson` and `Anonymous`
+* now `SaId`, `SaKeyId` and `SaPrivateKeyFile` have become deprecated - use `SaKeyJson` insted
 * added the ability to configure write attempt time
 * added graceful shutdown